--- a/.github/workflows/main.yml
+++ b/.github/workflows/main.yml
@@ -2,49 +2,70 @@ name: CI
 
 on:
   push:
-    branches:
-      - main
-      - "release/**"
 
-  pull_request:
-
+env:
+  PYTHON_VERSION: 3.9
+  DJANGO_SETTINGS_MODULE: config.settings.test
+  DATABASE_ENGINE: django.db.backends.postgresql
+  DATABASE_NAME: postgres
+  DATABASE_USER: postgres
+  DATABASE_PASSWORD: postgres
+  DATABASE_HOST: localhost
+  DATABASE_PORT: 5432
 jobs:
   lint-python:
     runs-on: ubuntu-latest
     steps:
-      - uses: actions/checkout@v2
-      - name: Set up Python
-        uses: actions/setup-python@v2
+      - uses: actions/checkout@v3
+      - uses: actions/setup-python@v4
         with:
-          python-version: 3.9
-      - name: Install flake8
-        run: python -m pip install flake8
+          python-version: ${{ env.PYTHON_VERSION }}
+          cache: 'pip'
+          cache-dependency-path: '**/lint-requirements.txt'
+      - run: pip install -r lint-requirements.txt
       - name: Check compliance with flake8
         run: flake8 .
-      - name: Install isort
-        run: python -m pip install isort
       - name: Check compliance with isort
         run: isort -c .
-      - name: Install Black
-        run: python -m pip install black
       - name: Check compliance with Black
         run: black etna config --check --fast
 
   test-python:
     runs-on: ubuntu-latest
     needs: [lint-python]
+    services:
+      postgres:
+        image: postgres:12.3
+        env:
+          POSTGRES_USER: postgres
+          POSTGRES_PASSWORD: postgres # pragma: allowlist secret
+          POSTGRES_DB: postgres
+        ports:
+          - 5432:5432
+        options: --health-cmd pg_isready --health-interval 1s --health-timeout 3s --health-retries 10
     steps:
-      - uses: actions/checkout@v2
-      - name: Create env file
-        run: cp .env.example .env
-      - name: Start containers
-        run: docker-compose up -d
-      - name: Run tests
-        run: |
-          docker-compose exec -T web poetry run ./manage.py test
+      - uses: actions/checkout@v3
+      - uses: actions/setup-python@v4
+        with:
+          python-version: ${{ env.PYTHON_VERSION }}
+      - uses: snok/install-poetry@v1
+        with:
+          virtualenvs-create: true
+          virtualenvs-in-project: true
+      - id: cached-poetry-dependencies
+        uses: actions/cache@v3
+        with:
+          path: .venv
+          key: venv-${{ runner.os }}-${{ hashFiles('**/poetry.lock') }}
+      - name: Install dependencies
+        if: steps.cached-poetry-dependencies.outputs.cache-hit != 'true'
+        run: poetry install --no-interaction --no-root
+      - name: Run Django checks
+        run: poetry run python manage.py check
+      - name: Run test suite
+        run: poetry run python manage.py test etna
       - name: Check for missing migrations
-        run: |
-          docker-compose exec -T web poetry run ./manage.py makemigrations --check
+        run: poetry run python manage.py makemigrations --check --noinput
 
   test-javascript:
     runs-on: ubuntu-latest