--- a/.github/workflows/main.yml
+++ b/.github/workflows/main.yml
@@ -2,6 +2,10 @@ name: CI
 
 on:
   push:
+    branches:
+      - main
+      - "release/**"
+  pull_request:
 
 env:
   PYTHON_VERSION: 3.9