--- a/lint-requirements.txt
+++ b/lint-requirements.txt
@@ -1,0 +1,3 @@
+flake8
+isort
+black