--- a/lint-requirements.txt
+++ b/lint-requirements.txt
@@ -1,3 +1,3 @@
-flake8
-isort
-black
+flake8==4.0.1
+isort==5.10.1
+black==22.3.0