… etna/categories/migrations/0003_alter_category_icon.py