With these changes: Average: ~110 seconds
- [1m 45s](https://github.com/nationalarchives/ds-wagtail/actions/runs/2555141191)
- [1m 44s](https://github.com/nationalarchives/ds-wagtail/actions/runs/2525127079)
- [1m 55s](https://github.com/nationalarchives/ds-wagtail/actions/runs/2525143643)

Before these changes. Average: ~267 seconds
- [3m 41s](https://github.com/nationalarchives/ds-wagtail/actions/runs/2523359963)
- [5m 2s](https://github.com/nationalarchives/ds-wagtail/actions/runs/2508254709)
- [4m 39s](https://github.com/nationalarchives/ds-wagtail/actions/runs/2487599941)

Improvement: 157 seconds (a 58% improvement)